pyproject.toml
[tool.poetry]
name = "social-network-inspiration"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "backend" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
django = "^5.1.6"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.7.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

